[#12]: Интегрировать линтер (#45)
* [+]: adding linter

* Update src/backend/internal/model/order.go


* [*]: fix naming error

---------

+++ src/backend/pkg/storage/postgres/core.go
@@ -16,7 +16,7 @@ const (
 )
 
 type Postgres struct {
-	maxPoolSize  int
+	maxPoolSize  int32
 	connAttempts int
 	connTimeout  time.Duration
 
@@ -42,7 +42,7 @@ func New(url string, opts ...Option) (*Postgres, error) {
 		return nil, fmt.Errorf("pgxpool parse config: %w", err)
 	}
 
-	poolConfig.MaxConns = int32(pg.maxPoolSize)
+	poolConfig.MaxConns = pg.maxPoolSize
 
 	for pg.connAttempts > 0 {
 		pg.Pool, err = pgxpool.NewWithConfig(context.Background(), poolConfig)
